Restructure layouts and add admin layout
